start splitting  interface between data an engine
