+++ src/engine.rs
@@ -39,3 +39,5 @@ mod journeys_tree;
 pub mod multicriteria_raptor;
 
 mod pareto_front;
+
+pub mod engine_interface;
