+++ src/engine/pareto_front.rs
@@ -35,7 +35,8 @@
 // www.navitia.io
 
 use crate::engine::journeys_tree::{Arrive, Board, Debark, Wait};
-use crate::traits::{Request, RequestTypes, TransitTypes};
+use crate::engine::engine_interface::{Request, RequestTypes};
+use crate::transit_data::data_interface::TransitTypes;
 
 use std::{fmt::Debug, slice::Iter as SliceIter};
