+++ src/request/generic_request.rs
@@ -41,10 +41,11 @@ use crate::{
 };
 
 use log::warn;
-use traits::{BadRequest, RequestTypes};
+use crate::engine::engine_interface::{BadRequest, RequestTypes, RequestInput};
 use transit_model::Model;
+use crate::transit_data::data_interface::Data as DataTrait;
 
-pub struct GenericRequest<'data, 'model, Data: traits::Data> {
+pub struct GenericRequest<'data, 'model, Data: DataTrait> {
     pub(super) transit_data: &'data Data,
     pub(super) model: &'model Model,
     pub(super) departure_datetime: SecondsSinceDatasetUTCStart,
@@ -59,12 +60,12 @@ pub struct GenericRequest<'data, 'model, Data: traits::Data> {
 
 impl<'data, 'model, Data> GenericRequest<'data, 'model, Data>
 where
-    Data: traits::Data,
+    Data: DataTrait,
 {
     pub fn new(
         model: &'model transit_model::Model,
         transit_data: &'data Data,
-        request_input: &traits::RequestInput,
+        request_input: &RequestInput,
     ) -> Result<Self, BadRequest>
     where
         Self: Sized,
@@ -158,10 +159,10 @@ where
 }
 
 use crate::response;
-use crate::traits::Journey as PTJourney;
+use crate::engine::engine_interface::Journey as PTJourney;
 impl<'data, 'model, Data> GenericRequest<'data, 'model, Data>
 where
-    Data: traits::Data,
+    Data: DataTrait,
 {
     pub fn create_response<R>(
         &self,
@@ -213,7 +214,6 @@ where
     }
 }
 
-use crate::traits::Data as DataTrait;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct Departure {
@@ -270,7 +270,7 @@ impl Iterator for Arrivals {
 
 impl<'data, 'model, Data> GenericRequest<'data, 'model, Data>
 where
-    Data: traits::Data,
+    Data: DataTrait,
 {
     pub fn stop_name(&self, stop: &Data::Stop) -> String {
         let stop_point_idx = self.transit_data.stop_point_idx(stop);
